+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Conrad Changelog
 
+## Version 1.0.0
+
+* Added ability to add a UUID to an event
+* **BREAKING CHANGE** Renamed middlewares to processors
+* **BREAKING CHANGE** Changed timestamp processor to accept keyword args
+* Added ability to change timestamp attribute key for timestamp processor
+
 ## Version 0.1.0
 
 * Added `StdoutEmitter`